Don't use link action feature configuration in LTO Backend command line
The link action feature configuration command line options aren't needed
in the LTO Backend compile command line. This additionally causes a failure
when trying to compile ThinLTO and FDO, as the build variables for
the FDO link options in the feature configuration are not set.

Confirmed that this is a no-op for non-FDO builds (there are no c++-link* features),
which is probably why I didn't notice this and remove it earlier. We add the
lto-backend features later in LTOBackendArtifacts.scheduleLTOBackendAction()

--
MOS_MIGRATED_REVID=134076174