Bazel bootstrapping: consume custom startup flags
The bootstrap build is special in that it does not
use the Bazel client hence we cannot pass a
--bazelrc flag as there'd be nothing to parse it.

While Bazel is being bootstrapped it's nice to
have colors on a smart terminal, but when the
bootstrapping is done by Jenkins as part of a CI
test, we need the output to be readable, i.e. be
free of control characters, so we should pass
--colors=no for example.

By exporting BAZEL_BOOTSTRAP_STARTUP_OPTIONS the
user can control what startup flags will the
bootstrap build use, so we can pass these
dumbing-down flags during CI builds.

--
MOS_MIGRATED_REVID=134273969