Make .d file analysis properly mark used module (.pcm) files.
Previously, we thought this wouldn't be necessary, as a change in .pcm would
always be accompanied by a corresponding change in a header. However, in the
current compile module this isn't true. Modules currently require all
transitive modules as inputs even though they might not use a single header
from them. Thus, if a module //a depends on a module //b, any change to //b's
headers will currently require //a to be rebuilt unless those headers are
actually included by //a. This leads to false positive cache results.

--
MOS_MIGRATED_REVID=135134348