Rollback of commit c1b9ae16283f3860eb1e3e4f6304133e8571ea6b.
*** Reason for rollback ***

With this approach, the R.class winds up in the filtered jar because its origin isn't among the "normal" Java sources. I'll have to be a little smarter, but start with rolling this back.

*** Original change description ***

Support srcjar files in filtered gen jar.

If a java_library has any srcjar files in its sources
attribute, the output of these would be discarded.

To make it more robust, invert the jar filter to
keep everything *except* source artifact java files.
Previously, we would build a manifest containing all
the generated java files and use that to filter.

This should also be ever so slightly faster since
we can skip an action.

--
MOS_MIGRATED_REVID=135406849