Rollback of commit ead1a810066e90cd8b822727de06c63ef942d3e7.
*** Reason for rollback ***

Depends on c/135226123 which depends on commit dd9344d1017fc2a8d72624a854b24dce10d66d17, which needs to be rolled
back.

*** Original change description ***

Add ctx.get_actions(), for inspecting the actions created by the current rule.

This returns an ActionsProvider. In the case where the rule does not emit any
more actions afterwards, the provider is equivalent to the one that gets passed
on to the rule's dependencies.

This may be useful for unit testing analysis-time helper functions that take
in ctx and have the side-effect of creating actions. In this use case, the
testing rule should be marked _skylark_testable=True, and its implementation
f...

***

--
MOS_MIGRATED_REVID=135781162