Add a design document on a distribution artifact for bazel
With bazel sources depending on checked-in binaries for the
supported platforms, porting bazel to a new platform is hard;
that approach also doesn't scale well if we want to support
users on more platforms.

This design document suggests an alternative approach based
on creating a zip-archive that, besides all sources, also
contains the generated output of the protoc actions. From
this archive, bazel can be bootstrapped without depending
on the right protoc on the target platform.

--
Change-Id: I401452435ed4189ea95260961d981ccc99a98560
Reviewed-on: https://bazel-review.googlesource.com/#/c/6530
MOS_MIGRATED_REVID=135891242