Automated [] rollback of commit 97e369465d51c2b4a7a10eb38a61e5e80796dec2.
*** Reason for rollback ***

Make darwin sandboxing default again.

*** Original change description ***

Fix #1849: Sandboxing on OS X should be turned off by default for 0.3.2.

This restructures the way we set the default Spawn strategy so that each BlazeModule supplying a SpawnActionContext has an ActionContextConsumer that sets its own SpawnActionContext as the default, with the BazelRulesModule being put as the last module loaded in BazelMain, so that it can override that decision - it only does, if the user explicitly specifies a --spawn_strategy flag. IMHO this is a much saner approach than...

***

ROLLBACK_OF=134770427

RELNOTES: Darwin sandboxing is default.

--
MOS_MIGRATED_REVID=135905657