Create a proper wrapper script for executing "bazel" in the integration tests.
Currently a call to "bazel" in an integration test means calling a (quite
hidden) function in test-setup.sh which actually calls "$bazel" defined
in "shell/bazel/testenv.sh" which is equal to "$(rlocation io_bazel/src/bazel)".
This is extremely confusing and error prone.

The new mechanism is to add a wrapper script to shell/bin called bazel
and export this directory to the PATH. 

Moreover, not every test loads the same test environment, for instance consider
how bazel_query_test loads the test environment:
- Load shell/integration/testenv.sh which loads,
- shell/bazel/test-setup.sh which loads,
- shell/bazel/testenv.sh which loads,
- shell/unittest.bash which loads,
- shell/testenv.sh

Again this is error prone and specially hard to understand, in fact
each test writer needs to decide which of these testenv to load. 
This change fixes all of this by having only one testenv.sh 
and summarizing the test setup in integration_test_setup.sh. 
Namely, for any new integration test, the developer
needs to load integration_test_setup to get the environment set up including
the unittest framework (also it helps to attract contributions).

This change also allows to open sourcing client_sigint_test: Since bazel was a 
function client_sigint_test was using a wrong process id to interrupt 
the build. The problem is that $! returns
bash's id instead of the id of the process running in the background 
when using a function instead of an executable.

A few tests needed to be adapted to the new infrastructure.

--
MOS_MIGRATED_REVID=136470360