Made HttpDownloader download calls non-static.
To set and use a RepositoryCache instance in HttpDownloader while parsing the command line options, we can pass an AtomicReference<HttpDownloader> instance from BazelRepositoryModule to the HttpArchiveFunctions. However, we'll need to change HttpDownloader download() calls to be non-static in order to initialize an instance of HttpDownloader in BazelRepositoryModule.

Remaining TODOs:

- RepositoryCache implementation and unit testing
- RepositoryCache lockfiles
- RepositoryCache integration testing

GITHUB: #1752

--
MOS_MIGRATED_REVID=136593517