Fix zip_manifest_creator dep in bazel_tools repo with an alias.
In commit 631d4f079eed6070472bb20a6370637ae0679bc6 I accidentally broke aar_import when I moved
zip_manifest_creator from tools/android to tools/zip because tools/android does
not have separate BUILD and BUILD.tools files. The test did not catch this,
because the dependency in tools/android/BUILD is correct in the context of the
bazel tree but not in the context of the bazel_tools embedded workspace.

--
MOS_MIGRATED_REVID=137209500