Initial implementation of "dylib" attribute for apple_binary and apple_dynamic_library rules
Provided values propagated from "dylib" dependencies will be compiled against the srcs of the rule, and linked together with the dependencies.
It is worth noting that "dylibs" differs from "deps" in that there is no configuration transition along this edge.

There is more work to be done on this attribute, so it remains undocumented. Namely, symbol deduping between dylib and statically-linked dependencies needs to be addressed.

--
MOS_MIGRATED_REVID=137721599