Do not tickle TimestampGranularityMonitor for CONSTANT_METADATA artifacts.
"Constant metadata" artifacts represent real files whose changes should be
ignored by the build system.  However, these artifacts were triggering the
timestamp granularity checks in TimestampGranularityMonitor because the fact
that they were "constant metadata" was not respected.  Avoid this so that
their regeneration does not cause the build to unnecessarily stall.

One of these artifacts is the volatile workspace status file, which is
unconditionally updated on each build.  Before this fix, "blaze build" would
get stuck for up to a second waiting for file system timestamps to catch up.
With this fix, the artifact is ignored and the wait is gone.  This problem
is magnified on macOS where the default HFS+ file system only has
second-level granularity.  (This also affects Linux, but because current
Linux file systems have milli/nanosecond-level granularity, the wait imposed
by TimestampGranularityMonitor is minimal and thus not generally noticeable.)

--
MOS_MIGRATED_REVID=137867586