On ci, bootstrap from a bootstrap bazel
With the implementation of the distribution-artifact design,
calling ./compile.sh on a git checkout won't work any more
unless you already have the correct version of the protoc
installed. To prepare for that, and to also keep the maintenance
burden for the ci slaves low, we build bazel on the ci machines
with the latest release of bazel as bootstrap bazel.

--
Change-Id: I18953489cc7282586d6d0a3911ec94e11215a540
Reviewed-on: https://bazel-review.googlesource.com/#/c/6771
MOS_MIGRATED_REVID=138761693