Adapt the installation instruction to the distfile
Change the installation instructions to be compatible with
the upcoming distribution artifact change. For compiling from
source, we recommend two ways,
- using the distribution archive and calling ./compile.sh, and
- using a bootstrap version of bazel already installed and calling
  bazel build //src:bazel.
The third way, of having the precise version of protoc needed already
installed is still possible but not recommended as it is very fragile.
So do not mention it.

--
Change-Id: I2f41913b37bcb0058a28e6b49c496611ae4b1908
Reviewed-on: https://bazel-review.googlesource.com/#/c/7132
MOS_MIGRATED_REVID=138871245