Fix bugs in slicing with a negative stride
There was a crash when the starting point of a negative slice is larger than
the list length.

In addition, there was an incorrect result when the ending point of a negative
slice is less than the list's length times -1. It would wrongly exclude the
first element.

RELNOTES: Fix slicing bug where "abc"[:-4:-1] would give wrong answer

--
MOS_MIGRATED_REVID=138878716