Changing constructor of JavaCompilationHelper and removing BaseJavaCompilationHelper.
Changed JavaCompilationHelper constructor to take in the object it used to construct itself: java toolchain, a host javabase and a jacoco instrumental. Because the objects are now directly passed to the constructor there is no more need anymore for the implicit attribute suffix, that was used to construct these objects.

I also removed BaseJavaCompilationHelper as it was only used to create an ijar action and only had one class extend it (JavaCompilationHelper). All methods that construct objects (java toolchain, host javabase, jacoco instrumental) were moved to JavaCompilationHelper as static methods and the rest of its method were moved as they were also to JavaCompilationHelper.

This is needed in order to pass the right toolchain/host javabase/jacoco instrumental from a Skylark module. The problem arises when a Skylark aspect/rule is created and its attributes will have different labels then a Java native rule/aspect because they must be prefixed with "_" instead of ":", so their value cannot be retrieved in the same default way.

--
MOS_MIGRATED_REVID=139310690