Use num files in directory instead of num links to file in testjobs_test
Adding and removing hard links to a file from several processes
simultaneously appears to occasionally fail on macOS.

This commit changes the testjobs's test's strategy of keeping track
of the number of tests running in parallel. Instead of counting the
number of links to a shared file, it now counts the number of files
in a shared directory. Instead of each test run adding and removing
a link to a shared file, each run adds and removes a file from a
directory.

--
Change-Id: Idc0a9a9cf1c2f5b03221fcd991e4bf0155e62841
Reviewed-on: https://bazel-review.googlesource.com/c/7392/
MOS_MIGRATED_REVID=139384412