Improve windows docs around C++
These are the options that a colleague needed to install in their VS 2015 Community Edition install before it was possible to compile C++ with Bazel. I've reflected this in the doc (I thought I remembered an open issue for this, but couldn't find it just now when I looked).

(I elected to not change the markdown formatting because of the `<br/>` in play. I think it's possible to get the same hanging indent by indenting the text within a bullet by one indent-width, but I have no way of previewing the rendering to know whether that's true or not.)

![cpp options in vs 2015](https://cloud.githubusercontent.com/assets/69720/20141432/3cfe71e6-a688-11e6-83cd-070fa60a54f2.PNG)

(Specifically, in the screenshot, I'm referring to the C++ ones, not any of the others).

Closes #2063.

--
Reviewed-on: https://github.com/bazelbuild/bazel/pull/2063
MOS_MIGRATED_REVID=140134095