Introducing the Java compilation black box in Skylark.
Adding a new declared provider that will be the returning value of the Java compilation method in Skylark. This provider is considered a black-box from Skylark and it should only be passed around (returned from a rule/aspect implementation) to allow Java rules to depend on it.

Also adding a new Skylark module (java_common) to allow retrieving the Java dependencies' black-box from Skylark implementations.

--
MOS_MIGRATED_REVID=140138446