Start removing flag for rclass generator (make nop)
Flip option default to true to mirror flag flip
and make the option a nop. After that we can remove
the flag reference, and then we can remove
the actual option.

RELNOTES: --{no}experimental_use_rclass_generator
is now a nop. It's been on by default for a while,
and there are no references, so seems safe to remove.

--
MOS_MIGRATED_REVID=140602921