Add --ltoindexopt to pass options to LTO indexing step
This makes it much easier to pass options just to the LTO indexing. Previously these had to be passed via --linkopt, but that also passed the options to the final native link, which causes issues since most useful ThinLTO indexing options are plugin options and gold gives an error when these are passed to native links not involving a plugin.

--
MOS_MIGRATED_REVID=140735846